+++ test/vhpi/vhpi10.c
@@ -234,7 +234,7 @@ static void start_of_sim(const vhpiCbDataT *cb_data)
    vhpi_printf("i0g1 type full name is %s", vhpi_get_str(vhpiFullNameP, i0g1_type));
    vhpi_printf("i0g1 dimensions %d", vhpi_get(vhpiNumDimensionsP, i0g1_type));
    fail_unless(vhpi_get(vhpiKindP, i0g1_type) == vhpiSubtypeDeclK);
-   fail_unless(vhpi_get(vhpiIsUnconstrainedP, i0g1_type));  // XXX: is this right?
+   fail_if(vhpi_get(vhpiIsUnconstrainedP, i0g1_type));
 
    vhpiHandleT i0g1_constrs = vhpi_iterator(vhpiConstraints, i0g1_type);
    check_error();