+++ NEWS.md
@@ -4,6 +4,8 @@
   `--load=siginit` and `+siginit` arguments. This is useful for finding
   reset issues (from @Blebowski).
 - Added support for System Verilog procedural assertions.
+- Fixed an issue where VHPI could not iterate the constraints of objects
+  declared with an unconstrained subtype (#1632).
 
 ## Version 1.22.1 - 2026-08-01
 - Waveform dumping now dumps multiple top units instead of only the first